qcacld-3.0: Add roam_reason in roam auth event
Currently, there is no information for roam reason
in hdd.

Fix, Handle roam reason recived from csr, print roam
reason and add roam reason in roam auth event.

CRs-Fixed: 2425910

+++ core/sme/inc/csr_api.h
@@ -1185,7 +1185,7 @@ struct csr_roam_info {
 	struct sir_sae_info *sae_info;
 #endif
 	struct assoc_ind *owe_pending_assoc_ind;
-	uint8_t roam_reason;
+	uint16_t roam_reason;
 };
 
 typedef struct sSirSmeAssocIndToUpperLayerCnf {
